+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 - MMark, extensions removed:
   - `ghcSyntaxHighlighter`: we already have `skylighting` (which supports more parsers than Haskell)
   - `obfuscateEmail`: requires JS, which is not documented.
+- API changes:
+  - Introduced `forEvery` to run a Shake action over a pattern of files when they change.
+  - Removed `buildHtmlMulti` (use `forEvery` with `buildHtml` instead)
+  - Exposed `Rib.Shake.writeFileCached`
 
 ## 0.6.0.0